chore: move macro in a submodule
