+++ crates/agglayer-storage/src/columns/local_exit_tree_per_network/mod.rs
@@ -33,7 +33,7 @@ pub enum Value {
     Frontier([u8; 32]),
 }
 
-impl_codec_using_bincode_for!(Key, Value);
+crate::columns::impl_codec_using_bincode_for!(Key, Value);
 
 impl ColumnSchema for LocalExitTreePerNetworkColumn {
     type Key = Key;
