+++ crates/agglayer-storage/src/columns/pending_queue/mod.rs
@@ -15,7 +15,7 @@ pub(crate) struct PendingQueueColumn;
 #[derive(Debug, Serialize, Deserialize)]
 pub struct PendingQueueKey(pub(crate) NetworkId, pub(crate) Height);
 
-impl_codec_using_bincode_for!(PendingQueueKey);
+crate::columns::impl_codec_using_bincode_for!(PendingQueueKey);
 
 impl ColumnSchema for PendingQueueColumn {
     type Key = PendingQueueKey;
