+++ crates/agglayer-storage/src/lib.rs
@@ -1,19 +1,3 @@
-macro_rules! impl_codec_using_bincode_for {
-    ($($type:ty),* $(,)?) => {
-        $(
-            impl $crate::columns::Codec for $type {
-                fn encode(&self) -> Result<Vec<u8>, $crate::columns::CodecError> {
-                    Ok($crate::columns::bincode_codec().serialize(self)?)
-                }
-
-                fn decode(buf: &[u8]) -> Result<Self, $crate::columns::CodecError> {
-                    Ok($crate::columns::bincode_codec().deserialize(buf)?)
-                }
-            }
-        )*
-    };
-}
-
 // Physical storage
 #[rustfmt::skip]
 pub mod storage;
